requirements.txt
alpaca-py>=0.8.0
python-dotenv>=0.19.0
